PKAlloc*: avoid overflow and code cleanup

+++ jxrgluelib/JXRGlue.c
@@ -2,16 +2,16 @@
 //
 // Copyright © Microsoft Corp.
 // All rights reserved.
-// 
+//
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are met:
-// 
+//
 // • Redistributions of source code must retain the above copyright notice,
 //   this list of conditions and the following disclaimer.
 // • Redistributions in binary form must reproduce the above copyright notice,
 //   this list of conditions and the following disclaimer in the documentation
 //   and/or other materials provided with the distribution.
-// 
+//
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 // AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 // IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -109,8 +109,7 @@ int PKStrnicmp(const char* s1, const char* s2, size_t c)
     return c ? *s1 - *s2 : 0;
 }
 
-static const PKPixelInfo pixelInfo[] =
-{
+static const PKPixelInfo pixelInfo[] = {
     {&GUID_PKPixelFormatDontCare, 1, Y_ONLY, BD_8, 8, PK_pixfmtNul, 0, 0, 0, 0},
 
     // Gray
@@ -181,7 +180,7 @@ static const PKPixelInfo pixelInfo[] =
     {&GUID_PKPixelFormat48bpp6Channels, 6, NCOMPONENT, BD_8, 48, PK_pixfmtNul, PK_PI_NCH, 6, 8, 1},
     {&GUID_PKPixelFormat56bpp7Channels, 7, NCOMPONENT, BD_8, 56, PK_pixfmtNul, PK_PI_NCH, 7, 8, 1},
     {&GUID_PKPixelFormat64bpp8Channels, 8, NCOMPONENT, BD_8, 64, PK_pixfmtNul, PK_PI_NCH, 8, 8, 1},
-    
+
     {&GUID_PKPixelFormat32bpp3ChannelsAlpha, 4, NCOMPONENT, BD_8, 32, PK_pixfmtHasAlpha, PK_PI_NCH, 4, 8, 1},
     {&GUID_PKPixelFormat40bpp4ChannelsAlpha, 5, NCOMPONENT, BD_8, 40, PK_pixfmtHasAlpha, PK_PI_NCH, 5, 8, 1},
     {&GUID_PKPixelFormat48bpp5ChannelsAlpha, 6, NCOMPONENT, BD_8, 48, PK_pixfmtHasAlpha, PK_PI_NCH, 6, 8, 1},
@@ -249,7 +248,7 @@ ERR PixelFormatLookup(PKPixelInfo* pPI, U8 uLookupType)
     Call(WMP_errUnsupportedFormat);
 
 Cleanup:
-    return err;        
+    return err;
 }
 
 
@@ -347,7 +346,7 @@ ERR PKCreateFactory_Release(PKFactory** ppFactory)
 
     Call(PKFree((void **) ppFactory));
 
-Cleanup: 
+Cleanup:
     return err;
 }
 
@@ -366,7 +365,7 @@ ERR PKCreateFactory(PKFactory** ppFactory, U32 uVersion)
 
     pFactory->CreateStreamFromFilename = CreateWS_File;
     pFactory->CreateStreamFromMemory = CreateWS_Memory;
-    
+
     pFactory->Release = PKCreateFactory_Release;
 
 Cleanup:
@@ -540,7 +539,7 @@ ERR PKImageEncode_SetSize(
 
 ERR PKImageEncode_SetResolution(
     PKImageEncode* pIE,
-    Float fResX, 
+    Float fResX,
     Float fResY)
 {
     pIE->fResX = fResX;
@@ -599,7 +598,7 @@ ERR PKImageEncode_WriteSource(
 
     U8* pb = NULL;
 
-	// CWMTranscodingParam* pParam = NULL; 
+    // CWMTranscodingParam* pParam = NULL;
 
     // get pixel format
     Call(pFC->GetSourcePixelFormat(pFC, &enPFFrom));
@@ -607,34 +606,34 @@ ERR PKImageEncode_WriteSource(
     FailIf(!IsEqualGUID(&pIE->guidPixFormat, &enPFTo), WMP_errUnsupportedFormat);
 
     // calc common stride
-//    Call(GetPixelInfo(enPFFrom, &pPIFrom));
+    //    Call(GetPixelInfo(enPFFrom, &pPIFrom));
     pPIFrom.pGUIDPixFmt = &enPFFrom;
     PixelFormatLookup(&pPIFrom, LOOKUP_FORWARD);
 
-//    Call(GetPixelInfo(enPFTo, &pPITo));
+    //    Call(GetPixelInfo(enPFTo, &pPITo));
     pPITo.pGUIDPixFmt = &enPFTo;
     PixelFormatLookup(&pPITo, LOOKUP_FORWARD);
 
-//    cbStrideFrom = (pPIFrom->cbPixel * pRect->Width + pPIFrom->cbPixelDenom - 1) / pPIFrom->cbPixelDenom;
-    cbStrideFrom = (BD_1 == pPIFrom.bdBitDepth ? ((pPIFrom.cbitUnit * pRect->Width + 7) >> 3) : (((pPIFrom.cbitUnit + 7) >> 3) * pRect->Width)); 
-    if (&GUID_PKPixelFormat12bppYUV420 == pPIFrom.pGUIDPixFmt 
-        || &GUID_PKPixelFormat16bppYUV422 == pPIFrom.pGUIDPixFmt) 
+    //    cbStrideFrom = (pPIFrom->cbPixel * pRect->Width + pPIFrom->cbPixelDenom - 1) / pPIFrom->cbPixelDenom;
+    cbStrideFrom = (BD_1 == pPIFrom.bdBitDepth ? ((pPIFrom.cbitUnit * pRect->Width + 7) >> 3) : (((pPIFrom.cbitUnit + 7) >> 3) * pRect->Width));
+    if (&GUID_PKPixelFormat12bppYUV420 == pPIFrom.pGUIDPixFmt
+        || &GUID_PKPixelFormat16bppYUV422 == pPIFrom.pGUIDPixFmt)
         cbStrideFrom >>= 1;
 
-//    cbStrideTo = (pPITo->cbPixel * pIE->uWidth + pPITo->cbPixelDenom - 1) / pPITo->cbPixelDenom;
-    cbStrideTo = (BD_1 == pPITo.bdBitDepth ? ((pPITo.cbitUnit * pIE->uWidth + 7) >> 3) : (((pPITo.cbitUnit + 7) >> 3) * pIE->uWidth)); 
+    //    cbStrideTo = (pPITo->cbPixel * pIE->uWidth + pPITo->cbPixelDenom - 1) / pPITo->cbPixelDenom;
+    cbStrideTo = (BD_1 == pPITo.bdBitDepth ? ((pPITo.cbitUnit * pIE->uWidth + 7) >> 3) : (((pPITo.cbitUnit + 7) >> 3) * pIE->uWidth));
     if (&GUID_PKPixelFormat12bppYUV420 == pPITo.pGUIDPixFmt
-        || &GUID_PKPixelFormat16bppYUV422 == pPITo.pGUIDPixFmt) 
+        || &GUID_PKPixelFormat16bppYUV422 == pPITo.pGUIDPixFmt)
         cbStrideTo >>= 1;
 
     cbStride = max(cbStrideFrom, cbStrideTo);
 
     // actual dec/enc with local buffer
-    Call(PKAllocAligned((void **) &pb, cbStride * pRect->Height, 128));
+    Call(PKAllocAligned((void **) &pb, (size_t)cbStride * pRect->Height, 128));
 
     Call(pFC->Copy(pFC, pRect, pb, cbStride));
 
-	Call(pIE->WritePixels(pIE, pRect->Height, pb, cbStride));
+    Call(pIE->WritePixels(pIE, pRect->Height, pb, cbStride));
 
 Cleanup:
     PKFreeAligned((void **) &pb);
@@ -684,7 +683,7 @@ ERR PKImageEncode_Transcode(
 
     U8* pb = NULL;
 
-    CWMTranscodingParam cParam = {0}; 
+    CWMTranscodingParam cParam = {0};
 
     // get pixel format
     Call(pFC->GetSourcePixelFormat(pFC, &enPFFrom));
@@ -692,24 +691,24 @@ ERR PKImageEncode_Transcode(
     FailIf(!IsEqualGUID(&pIE->guidPixFormat, &enPFTo), WMP_errUnsupportedFormat);
 
     // calc common stride
-//    Call(GetPixelInfo(enPFFrom, &pPIFrom));
+    //    Call(GetPixelInfo(enPFFrom, &pPIFrom));
     pPIFrom.pGUIDPixFmt = &enPFFrom;
     PixelFormatLookup(&pPIFrom, LOOKUP_FORWARD);
 
-//    Call(GetPixelInfo(enPFTo, &pPITo));
+    //    Call(GetPixelInfo(enPFTo, &pPITo));
     pPITo.pGUIDPixFmt = &enPFTo;
     PixelFormatLookup(&pPITo, LOOKUP_FORWARD);
 
-//    cbStrideFrom = (pPIFrom->cbPixel * pRect->Width + pPIFrom->cbPixelDenom - 1) / pPIFrom->cbPixelDenom;
-    cbStrideFrom = (BD_1 == pPIFrom.bdBitDepth ? ((pPIFrom.cbitUnit * pRect->Width + 7) >> 3) : (((pPIFrom.cbitUnit + 7) >> 3) * pRect->Width)); 
-    if (&GUID_PKPixelFormat12bppYUV420 == pPIFrom.pGUIDPixFmt 
-        || &GUID_PKPixelFormat16bppYUV422 == pPIFrom.pGUIDPixFmt) 
+    //    cbStrideFrom = (pPIFrom->cbPixel * pRect->Width + pPIFrom->cbPixelDenom - 1) / pPIFrom->cbPixelDenom;
+    cbStrideFrom = (BD_1 == pPIFrom.bdBitDepth ? ((pPIFrom.cbitUnit * pRect->Width + 7) >> 3) : (((pPIFrom.cbitUnit + 7) >> 3) * pRect->Width));
+    if (&GUID_PKPixelFormat12bppYUV420 == pPIFrom.pGUIDPixFmt
+        || &GUID_PKPixelFormat16bppYUV422 == pPIFrom.pGUIDPixFmt)
         cbStrideFrom >>= 1;
 
-//    cbStrideTo = (pPITo->cbPixel * pIE->uWidth + pPITo->cbPixelDenom - 1) / pPITo->cbPixelDenom;
-    cbStrideTo = (BD_1 == pPITo.bdBitDepth ? ((pPITo.cbitUnit * pIE->uWidth + 7) >> 3) : (((pPITo.cbitUnit + 7) >> 3) * pIE->uWidth)); 
+    //    cbStrideTo = (pPITo->cbPixel * pIE->uWidth + pPITo->cbPixelDenom - 1) / pPITo->cbPixelDenom;
+    cbStrideTo = (BD_1 == pPITo.bdBitDepth ? ((pPITo.cbitUnit * pIE->uWidth + 7) >> 3) : (((pPITo.cbitUnit + 7) >> 3) * pIE->uWidth));
     if (&GUID_PKPixelFormat12bppYUV420 == pPITo.pGUIDPixFmt
-        || &GUID_PKPixelFormat16bppYUV422 == pPITo.pGUIDPixFmt) 
+        || &GUID_PKPixelFormat16bppYUV422 == pPITo.pGUIDPixFmt)
         cbStrideTo >>= 1;
 
     cbStride = max(cbStrideFrom, cbStrideTo);
@@ -720,21 +719,21 @@ ERR PKImageEncode_Transcode(
         cParam.cWidth = pFC->pDecoder->WMP.wmiI.cROIWidth;
         cParam.cHeight = pFC->pDecoder->WMP.wmiI.cROIHeight;
         cParam.oOrientation = pFC->pDecoder->WMP.wmiI.oOrientation;
-//        cParam.cfColorFormat = pFC->pDecoder->WMP.wmiI.cfColorFormat;
+        //        cParam.cfColorFormat = pFC->pDecoder->WMP.wmiI.cfColorFormat;
         cParam.uAlphaMode = pFC->pDecoder->WMP.wmiSCP.uAlphaMode;
         cParam.bfBitstreamFormat = pFC->pDecoder->WMP.wmiSCP.bfBitstreamFormat;
         cParam.sbSubband = pFC->pDecoder->WMP.wmiSCP.sbSubband;
         cParam.bIgnoreOverlap = pFC->pDecoder->WMP.bIgnoreOverlap;
-        
+
         Call(pIE->Transcode(pIE, pFC->pDecoder, &cParam));
     }
-	else 
-	{
-		// actual dec/enc with local buffer
-	    Call(PKAllocAligned((void **) &pb, cbStride * pRect->Height, 128));
-		Call(pFC->Copy(pFC, pRect, pb, cbStride));
-		Call(pIE->WritePixels(pIE, pRect->Height, pb, cbStride));
-	}
+    else
+    {
+        // actual dec/enc with local buffer
+        Call(PKAllocAligned((void **) &pb, (size_t)cbStride * pRect->Height, 128));
+        Call(pFC->Copy(pFC, pRect, pb, cbStride));
+        Call(pIE->WritePixels(pIE, pRect->Height, pb, cbStride));
+    }
 
 Cleanup:
     PKFreeAligned((void **) &pb);
@@ -778,7 +777,7 @@ ERR PKImageEncode_Create(PKImageEncode** ppIE)
     pIE->SetColorContext = PKImageEncode_SetColorContext;
     pIE->SetDescriptiveMetadata = PKImageEncode_SetDescriptiveMetadata;
     pIE->WritePixels = PKImageEncode_WritePixels;
-//    pIE->WriteSource = PKImageEncode_WriteSource;
+    //    pIE->WriteSource = PKImageEncode_WriteSource;
 
     pIE->WritePixelsBandedBegin = PKImageEncode_WritePixelsBandedBegin;
     pIE->WritePixelsBanded = PKImageEncode_WritePixelsBanded;
@@ -786,12 +785,12 @@ ERR PKImageEncode_Create(PKImageEncode** ppIE)
 
     pIE->CreateNewFrame = PKImageEncode_CreateNewFrame;
     pIE->Release = PKImageEncode_Release;
-	pIE->bWMP = FALSE; 
+    pIE->bWMP = FALSE;
 
 Cleanup:
     return err;
 }
-  
+
 
 //================================================================
 // PKImageDecode
